+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add appengine group subcommand
 - Add support for credentials inhibition
 - Add support for interfaces stats in device describe subcommand
+- Add appengine stats subcommand
 
 ## [0.10.3] - 2019-11-06
 ### Added